fix tests, reduce diff
